Move runtime.interfaces to runtime.name
